Fixed Code Smells
- Removed all code smells in timesheet project

+++ src/main/java/tn/esprit/spring/repository/DepartementRepository.java
@@ -1,11 +1,13 @@
 package tn.esprit.spring.repository;
 
+
+
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import tn.esprit.spring.entities.Departement;
-
+import tn.esprit.spring.entity.Departement;
 @Repository
-public interface DepartementRepository extends CrudRepository<Departement, Integer>{
-
+public interface DepartementRepository extends CrudRepository<Departement, Long>{
+	
+	
 }
